Addition of privacy policy to menu
To comply with privacy policy regulations for apps on the Play Store, we must link to and include a privacy policy with the app.
This commit contains the addition of a 'Privacy Policy' option to the menu (accessible from upper right hand corner), along with moving the language options to their own 'Language' sub menu.

This commit also contains some other minor refactoring changes.